requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=4.5.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
Pillow>=10.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.0
drf-spectacular>=0.26.0
django-filter>=23.0
